+++ elements/pfe-accordion/pfe-accordion-panel.ts
@@ -1,3 +1,5 @@
+import type { ColorTheme } from '@patternfly/pfe-core';
+
 import { LitElement, html } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 
@@ -7,7 +9,6 @@ import { getRandomId } from '@patternfly/pfe-core/functions/random.js';
 import { PfeCollapsePanel } from '@patternfly/pfe-collapse/pfe-collapse-panel.js';
 
 import style from './pfe-accordion-panel.scss';
-import { ColorTheme } from '@patternfly/pfe-core';
 
 /**
  * Accordion Panel