+++ elements/pfe-accordion/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @patternfly/pfe-accordion
 
+## 2.0.0-next.5
+
+### Patch Changes
+
+- bfad8b4b: Updates dependencies
+- Updated dependencies [bfad8b4b]
+- Updated dependencies [a423b010]
+  - @patternfly/pfe-core@2.0.0-next.8
+  - @patternfly/pfe-collapse@2.0.0-next.3
+  - @patternfly/pfe-icon@2.0.0-next.4
+
 ## 2.0.0-next.4
 
 ### Major Changes