+++ tools/pfe-tools/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @patternfly/pfe-tools
 
+## 1.0.0-next.29
+
+### Minor Changes
+
+- 3f0c6ca2: Adds an optional unified config file for custom elements manifest, dev server, and docs pages
+
+  Create a `.pfe.config.json` file at the root of your project to customize the pfe tools builds.
+
+  The default values are:
+
+  ```json
+  {
+    "tagPrefix": "pfe",
+    "demoURLPrefix": "https://patternflyelements.org/",
+    "sourceControlURLPrefix": "https://github.com/patternfly/patternfly-elements/tree/main/",
+    "aliases": {},
+    "site": {
+      "title": "PatternFly Elements",
+      "description": "PatternFly Elements: A set of community-created web components based on PatternFly design.",
+      "favicon": "/brand/logo/svg/pfe-icon-blue.svg",
+      "logoUrl": "/brand/logo/svg/pfe-icon-white-shaded.svg",
+      "stylesheets": []
+    }
+  }
+  ```
+
+  See `@patternfly/pfe-tools/config.d.ts` for more information.
+
+## 1.0.0-next.28
+
+### Patch Changes
+
+- bfad8b4b: Updates dependencies
+- bfad8b4b: fix local module resolution in dev server
+- a423b010: fix typescript config and update dependencies
+- bfad8b4b: dev server: Resolve local elements from their source files
+
+## 1.0.0-next.27
+
+### Patch Changes
+
+- c625b329: Store nav state
+- a44b5484: Serve changes to TypeScript sources
+
 ## 1.0.0-next.26
 
 ### Patch Changes